chore(webview): show a button from VS Code UI Toolkit
Also moves things a bit more to where they belong
