+++ esbuild.js
@@ -3,6 +3,7 @@
 
 const { build } = require('esbuild')
 
+/** @type {import('esbuild').BuildOptions} */
 const baseConfig = {
   bundle: true,
   minify: process.env.NODE_ENV === 'production',
@@ -21,9 +22,19 @@ const extensionConfig = {
   external: ['vscode'],
 }
 
+/** @type {import('esbuild').BuildOptions} */
+const webviewConfig = {
+  ...baseConfig,
+  target: 'es2020',
+  format: 'esm',
+  entryPoints: ['./src/webview/main.ts'],
+  outfile: './dist/webview.js',
+}
+
 ;(async () => {
   try {
     await build(extensionConfig)
+    await build(webviewConfig)
     console.log('build complete')
   } catch (err) {
     process.stderr.write(err.stderr)
