+++ src/helpers/index.ts
@@ -6,3 +6,4 @@ export * from './fileWatcher'
 export * from './logExtensionActivated'
 export * from './radCli'
 export * from './views'
+export * from './webview'
