+++ src/utils/index.ts
@@ -1,6 +1,8 @@
 export * from './askUser'
 export * from './assertUnreachable'
 export * from './exec'
+export * from './getNonce'
+export * from './getUri'
 export * from './git'
 export * from './log'
 export * from './memoizeWithDebouncedCacheClear'
