+++ src/webview/main.ts
@@ -1,0 +1,3 @@
+import { provideVSCodeDesignSystem, vsCodeButton } from '@vscode/webview-ui-toolkit'
+
+provideVSCodeDesignSystem().register(vsCodeButton())
